Only propagate code based call annotations
