+++ mythril/laser/ethereum/plugins/implementations/mutation_pruner.py
@@ -1,24 +1,14 @@
-from mythril.laser.ethereum.state.annotation import StateAnnotation
-from mythril.laser.ethereum.svm import LaserEVM
 from mythril.laser.ethereum.plugins.signals import PluginSkipWorldState
 from mythril.laser.ethereum.plugins.plugin import LaserPlugin
+from mythril.laser.ethereum.plugins.implementations.plugin_annotations import (
+    MutationAnnotation,
+)
 from mythril.laser.ethereum.state.global_state import GlobalState
+from mythril.laser.ethereum.svm import LaserEVM
 from mythril.laser.ethereum.transaction.transaction_models import (
     ContractCreationTransaction,
 )
 from mythril.laser.smt import And, symbol_factory
-
-
-class MutationAnnotation(StateAnnotation):
-    """Mutation Annotation
-
-    This is the annotation used by the MutationPruner plugin to record mutations
-    """
-
-    @property
-    def persist_to_world_state(self):
-        # This should persist among calls, and be but as a world state annotation.
-        return True
 
 
 class MutationPruner(LaserPlugin):
@@ -47,16 +37,10 @@ class MutationPruner(LaserPlugin):
         @symbolic_vm.pre_hook("SSTORE")
         def sstore_mutator_hook(global_state: GlobalState):
             global_state.annotate(MutationAnnotation())
-            assert len(
-                list(global_state.world_state.get_annotations(MutationAnnotation))
-            )
 
         @symbolic_vm.pre_hook("CALL")
         def call_mutator_hook(global_state: GlobalState):
             global_state.annotate(MutationAnnotation())
-            assert len(
-                list(global_state.world_state.get_annotations(MutationAnnotation))
-            )
 
         @symbolic_vm.laser_hook("add_world_state")
         def world_state_filter_hook(global_state: GlobalState):
@@ -72,8 +56,5 @@ class MutationPruner(LaserPlugin):
                 global_state.current_transaction, ContractCreationTransaction
             ):
                 return
-            if (
-                len(list(global_state.world_state.get_annotations(MutationAnnotation)))
-                == 0
-            ):
+            if len(list(global_state.get_annotations(MutationAnnotation))) == 0:
                 raise PluginSkipWorldState
